put more things into machine-specific files, clean irrelevant function arguments
