+++ Tools/performance_tests/summit.py
@@ -1,4 +1,5 @@
-import os
+import os, copy
+from functions_perftest import test_element
 
 def executable_name(compiler,architecture):
     return 'perf_tests3d.' + compiler + '.TPROF.MPI.ACC.CUDA.ex'
@@ -48,6 +49,9 @@ def process_analysis(automated, cwd, compiler, architecture, n_node_list, start_
 def time_min(nb_simulations):
     return 2. + nb_simulations*2.
 
+def get_submit_job_command():
+    return ' bsub '
+
 def get_batch_string(test_list, job_time_min, Cname, n_node):
 
     job_time_str = str(int(job_time_min/60)) + ':' + str(int(job_time_min%60))
@@ -63,7 +67,7 @@ def get_batch_string(test_list, job_time_min, Cname, n_node):
     batch_string += 'module load cuda\n' 
     return batch_string
 
-def get_run_string(current_test, architecture, n_node, count, bin_name, runtime_param_list):
+def get_run_string(current_test, architecture, n_node, count, bin_name, runtime_param_string):
 
     output_filename = 'out_' + '_'.join([current_test.input_file, str(n_node), str(current_test.n_mpi_per_node), str(current_test.n_omp), str(count)]) + '.txt'
 
@@ -74,7 +78,7 @@ def get_run_string(current_test, architecture, n_node, count, bin_name, runtime_
     srun_string += ' -a ' + ngpu + ' -g ' + ngpu + ' -c ' + ngpu + ' --bind=packed:1 '
     srun_string += ' ./' + bin_name + ' '
     srun_string += current_test.input_file + ' '
-    srun_string += runtime_param_list[ count ]
+    srun_string += runtime_param_string
     srun_string += ' > ' + output_filename + '\n'
     return srun_string
